requirements.txt
selenium==4.0.0
Pillow>=10.1
beautifulsoup4==4.12.3
soupsieve==2.5
lxml==5.2.1
charset-normalizer==3.3.2
requests==2.31.0
